Some more adjustments and hide when sleeping
