+++ src/main/resources/assets/minecolonies/gui/townhall/windowtownhall.xml
@@ -6,9 +6,8 @@
     <image id="infopage0" pos="56 97" size="31 14" source="minecolonies:textures/gui/bookmark_short_ribbon_01.png"/>
     <image id="permissions0" pos="56 121" size="31 14" source="minecolonies:textures/gui/bookmark_short_ribbon_03.png"/>
     <image id="citizens0" pos="56 145" size="31 14" source="minecolonies:textures/gui/bookmark_short_ribbon_04.png"/>
-    <image id="workOrder0" pos="56 169" size="31 14" source="minecolonies:textures/gui/bookmark_short_ribbon_06.png"/>
-    <image id="happiness0" pos="56 193" size="31 14" source="minecolonies:textures/gui/bookmark_short_ribbon_01.png"/>
-    <image id="alliances0" pos="56 219" size="31 14" source="minecolonies:textures/gui/bookmark_short_ribbon_01.png"/>
+    <image id="happiness0" pos="56 169" size="31 14" source="minecolonies:textures/gui/bookmark_short_ribbon_01.png"/>
+    <image id="alliances0" pos="56 193" size="31 14" source="minecolonies:textures/gui/bookmark_short_ribbon_01.png"/>
 
     <buttonimage id="actions1" size="204 17" pos="58 43" source="minecolonies:textures/gui/bookmark_ribbon_02.png" visible="false"
                  label="$(com.minecolonies.coremod.gui.townhalltab.actions)"
@@ -21,9 +20,6 @@
                  textalign="MIDDLE_LEFT" textoffset="35 2" textcolor="White" textdisabledcolor="DarkGray"/>
     <buttonimage id="citizens1" size="204 17" pos="58 43" source="minecolonies:textures/gui/bookmark_ribbon_04.png" visible="false"
                  label="$(com.minecolonies.coremod.gui.townhalltab.citizens)"
-                 textalign="MIDDLE_LEFT" textoffset="35 2" textcolor="White" textdisabledcolor="DarkGray"/>
-    <buttonimage id="workOrder1" size="204 17" pos="58 43" source="minecolonies:textures/gui/bookmark_ribbon_06.png" visible="false"
-                 label="$(com.minecolonies.coremod.gui.townhalltab.workorders)"
                  textalign="MIDDLE_LEFT" textoffset="35 2" textcolor="White" textdisabledcolor="DarkGray"/>
     <buttonimage id="happiness1" size="204 17" pos="58 43" source="minecolonies:textures/gui/bookmark_ribbon_01.png" visible="false"
                  label="$(com.minecolonies.coremod.gui.townhalltab.stats)"
@@ -44,28 +40,23 @@
     <buttonimage id="citizensExt" pos="-17 145" size="104 14" source="minecolonies:textures/gui/bookmark_medium_ribbon_04.png" visible="false"
                  label="$(com.minecolonies.coremod.gui.townhalltab.citizens)"
                  textalign="MIDDLE_LEFT" textcolor="White" textoffset="8 1"/>
-    <buttonimage id="workOrderExt" pos="-17 169" size="104 14" source="minecolonies:textures/gui/bookmark_medium_ribbon_06.png" visible="false"
-                 label="$(com.minecolonies.coremod.gui.townhalltab.workorders)"
-                 textalign="MIDDLE_LEFT" textcolor="White" textoffset="8 1"/>
-    <buttonimage id="happinessExt" pos="-17 193" size="104 14" source="minecolonies:textures/gui/bookmark_medium_ribbon_01.png" visible="false"
+    <buttonimage id="happinessExt" pos="-17 169" size="104 14" source="minecolonies:textures/gui/bookmark_medium_ribbon_01.png" visible="false"
                  label="$(com.minecolonies.coremod.gui.townhalltab.stats)"
                  textalign="MIDDLE_LEFT" textcolor="White" textoffset="8 1"/>
-    <buttonimage id="alliancesExt" pos="-17 219" size="104 14" source="minecolonies:textures/gui/bookmark_medium_ribbon_01.png" visible="false"
+    <buttonimage id="alliancesExt" pos="-17 193" size="104 14" source="minecolonies:textures/gui/bookmark_medium_ribbon_01.png" visible="false"
                  label="$(com.minecolonies.coremod.gui.townhalltab.alliances)"
                  textalign="MIDDLE_LEFT" textcolor="White" textoffset="8 1"/>
 
     <buttonimage id="actions" size="17 17" pos="62 71" onHoverId="actionsExt"
                  source="minecolonies:textures/gui/red_wax_home.png"/>
     <buttonimage id="infopage" size="17 17" pos="62 95" onHoverId="infoExt"
-                 source="minecolonies:textures/gui/red_wax_information.png"/>
+                 source="minecolonies:textures/gui/red_wax_work_orders.png"/>
     <buttonimage id="permissions" size="17 17" pos="62 119" onHoverId="permissionsExt"
                  source="minecolonies:textures/gui/red_wax_permissions.png"/>
     <buttonimage id="citizens" size="17 17" pos="62 143" onHoverId="citizensExt"
                  source="minecolonies:textures/gui/red_wax_citizens.png"/>
-    <buttonimage id="workOrder" size="17 17" pos="62 167" onHoverId="workOrderExt"
-                 source="minecolonies:textures/gui/red_wax_work_orders.png"/>
-    <buttonimage id="happiness" size="17 17" pos="62 191" onHoverId="happinessExt"
+    <buttonimage id="happiness" size="17 17" pos="62 167" onHoverId="happinessExt"
                  source="minecolonies:textures/gui/red_wax_stats.png"/>
-    <buttonimage id="alliances" size="17 17" pos="62 215" onHoverId="alliancesExt"
-                 source="minecolonies:textures/gui/red_wax_stats.png"/>
+    <buttonimage id="alliances" size="17 17" pos="62 191" onHoverId="alliancesExt"
+                 source="minecolonies:textures/gui/red_wax_information.png"/>
 </window>
